+++ client/src/components/loader.jsx
@@ -1,0 +1,10 @@
+const Loader = () => (
+  <section className="dots-container">
+    <div className="dot"></div>
+    <div className="dot"></div>
+    <div className="dot"></div>
+    <div className="dot"></div>
+    <div className="dot"></div>
+  </section>
+);
+export default Loader;
